backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
openai==1.58.1
pdf2image==1.17.0
Pillow==11.0.0
weasyprint==63.1
python-dotenv==1.0.1
pydantic==2.10.3
